+++ src/main/java/com/teamtea/eclipticseasons/api/misc/CustomRandomTick.java
@@ -9,4 +9,6 @@ public interface CustomRandomTick {
     void eclipticseasons$tick(BlockState state, ServerLevel worldIn, BlockPos pos);
 
     void eclipticseasons$reset();
+
+    int eclipticseasons$getType();
 }