+++ src/main/java/com/teamtea/eclipticseasons/common/registry/ESRegistries.java
@@ -8,6 +8,7 @@ import com.teamtea.eclipticseasons.api.data.craft.WetterStructure;
 import com.teamtea.eclipticseasons.api.data.crop.CropGrowControlBuilder;
 import com.teamtea.eclipticseasons.api.data.quest.SeasonQuest;
 import com.teamtea.eclipticseasons.api.data.season.SeasonCycle;
+import com.teamtea.eclipticseasons.api.data.season.SeasonDefinition;
 import com.teamtea.eclipticseasons.api.data.season.SeasonPhase;
 import com.teamtea.eclipticseasons.api.data.season.SnowDefinition;
 import com.teamtea.eclipticseasons.api.data.weather.CustomRainBuilder;
@@ -43,6 +44,8 @@ public class ESRegistries {
     public static final ResourceKey<Registry<CustomSnowTerm>> SNOW_TERM = ResourceKey.createRegistryKey(EclipticSeasons.rl("snow_term"));
 
     public static final ResourceKey<Registry<WeatherRegion>> WEATHER_REGION = ResourceKey.createRegistryKey(EclipticSeasons.rl("weather_region"));
+
+    public static final ResourceKey<Registry<SeasonDefinition>> SEASON_DEFINITION = ResourceKey.createRegistryKey(EclipticSeasons.rl("season_definitions"));
 
     public static <T> String createLangKey(ResourceKey<Registry<T>> registryResourceKey, ResourceLocation resourceLocation) {
         // return resourceLocation.toLanguageKey(registryResourceKey.location().getPath());