+++ src/main/java/com/teamtea/eclipticseasons/data/extend/example/DatapackRegistryGeneratorExample.java
@@ -18,6 +18,7 @@ public class DatapackRegistryGeneratorExample extends DatapackBuiltinEntriesProv
             .add(ESRegistries.SNOW_TERM, SnowTermRegistry::bootstrap2)
             .add(ESRegistries.CROP, CropRegistry::bootstrap2)
             .add(ESRegistries.WEATHER_REGION, WeatherRegionRegistry::bootstrap2)
+            .add(ESRegistries.SEASON_DEFINITION, SeasonDefinitionRegistry::bootstrap2)
             ;
 
     public DatapackRegistryGeneratorExample(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {