+++ src/main/java/com/teamtea/eclipticseasons/mixin/data/RegistrySetBuilderMixin.java
@@ -17,10 +17,13 @@ public abstract class RegistrySetBuilderMixin {
     private Set<ResourceKey<Object>> eclipticseasons$buildPatch$fixError(Set<ResourceKey<Object>> original) {
         // because we not generate feature by dg, so it is not registered.
         // if ((Object) this instanceof DatapackRegistryGenerator)
-        if (original.contains(Biomes.PLAINS) || original.contains(Biomes.SUNFLOWER_PLAINS)) {
+        if (original.contains(Biomes.PLAINS)
+                || original.contains(Biomes.SUNFLOWER_PLAINS)
+                || original.contains(Biomes.THE_VOID)) {
             HashSet<ResourceKey<Object>> resourceKeys = new HashSet<>(original);
             resourceKeys.remove(Biomes.PLAINS);
             resourceKeys.remove(Biomes.SUNFLOWER_PLAINS);
+            resourceKeys.remove(Biomes.THE_VOID);
             return resourceKeys;
         }
         return original;